Tools: Allow pre-packaged job.

+++ tools/submit-job-to-k8s.sh
@@ -7,6 +7,7 @@
 
 # Default value for configurable arguments.
 JOB_FILE=""
+FUELING_PKG=""
 IMAGE="hub.baidubce.com/apollo/spark:latest"
 CONDA_ENV="fuel-py27-cyber"
 EXECUTORS=8
@@ -48,6 +49,10 @@ while [ $# -gt 0 ]; do
       shift
       EXECUTOR_DISK_GB=$1
       ;;
+    --fueling)
+      shift
+      FUELING_PKG=$1
+      ;;
     --partner_bos_region)
       shift
       PARTNER_BOS_REGION=$1
@@ -65,14 +70,10 @@ while [ $# -gt 0 ]; do
       PARTNER_BOS_SECRET=$1
       ;;
     *)
-      if [ -f "$1" ]; then
-        JOB_FILE=$1
-        shift
-        break
-      else
-        echo -e "$1: Unknown option or file not exists."
-        exit 1
-      fi
+      JOB_FILE=$1
+      echo "Spawning job: ${JOB_FILE}"
+      shift
+      break
       ;;
   esac
   shift
@@ -88,6 +89,8 @@ K8S="https://180.76.98.43:6443"
 DRIVER_MEMORY=2g
 FUEL_PATH="$( dirname "${BASH_SOURCE[0]}" )/.."
 BOS_FSTOOL_EXECUTABLE="${FUEL_PATH}/apps/local/bos_fstool"
+BOS_MOUNT_PATH="/mnt/bos"
+EVENTS_LOG_PATH=${BOS_MOUNT_PATH}/modules/data/spark/spark-events
 # End of config.
 
 set -x
@@ -102,28 +105,30 @@ fi
 
 set -e
 
-# Upload local files to remote.
-BOS_MOUNT_PATH="/mnt/bos"
-REMOTE_JOB_PATH="modules/data/jobs/$(date +%Y%m%d-%H%M%S)_${USER}"
-REMOTE_JOB_FILE="${REMOTE_JOB_PATH}/$(basename ${JOB_FILE})"
-REMOTE_FUELING_PKG="${REMOTE_JOB_PATH}/fueling.zip"
-
-"${BOS_FSTOOL_EXECUTABLE}" -s "${JOB_FILE}" -d "${REMOTE_JOB_FILE}"
-REMOTE_JOB_FILE="${BOS_MOUNT_PATH}/${REMOTE_JOB_FILE}"
-
-EVENTS_LOG_PATH=${BOS_MOUNT_PATH}/modules/data/spark/spark-events
-
-pushd "$( dirname "${BASH_SOURCE[0]}" )/.."
-  LOCAL_FUELING_PKG="deploy/fueling.zip"
-  rm -f "${LOCAL_FUELING_PKG}"
-  zip -r "${LOCAL_FUELING_PKG}" ./fueling -x *.pyc */__pycache__
-  "${BOS_FSTOOL_EXECUTABLE}" -s "${LOCAL_FUELING_PKG}" -d "${REMOTE_FUELING_PKG}" 
-  REMOTE_FUELING_PKG="${BOS_MOUNT_PATH}/${REMOTE_FUELING_PKG}"
-popd
-
 # Add kubernetes package to the spark-submit tool.
 rsync -aht --size-only "${FUEL_PATH}/apps/local/spark-kubernetes_2.11-2.4.0.jar" \
     ${CONDA_PREFIX}/lib/python3.6/site-packages/pyspark/jars/
+
+REMOTE_JOB_PATH="modules/data/jobs/$(date +%Y%m%d-%H%M%S)_${USER}"
+if [ -f ${JOB_FILE} ]; then
+  # Upload local job file to remote.
+  REMOTE_JOB_FILE="${REMOTE_JOB_PATH}/$(basename ${JOB_FILE})"
+  "${BOS_FSTOOL_EXECUTABLE}" -s "${JOB_FILE}" -d "${REMOTE_JOB_FILE}"
+  JOB_FILE="${BOS_MOUNT_PATH}/${REMOTE_JOB_FILE}"
+fi
+
+if [ -z ${FUELING_PKG} ]; then
+  # Upload local fueling package to remote.
+  REMOTE_FUELING_PKG="${REMOTE_JOB_PATH}/fueling.zip"
+
+  pushd "$( dirname "${BASH_SOURCE[0]}" )/.."
+    LOCAL_FUELING_PKG="deploy/fueling.zip"
+    rm -f "${LOCAL_FUELING_PKG}"
+    zip -r "${LOCAL_FUELING_PKG}" ./fueling -x *.pyc */__pycache__
+    "${BOS_FSTOOL_EXECUTABLE}" -s "${LOCAL_FUELING_PKG}" -d "${REMOTE_FUELING_PKG}" 
+    FUELING_PKG="${BOS_MOUNT_PATH}/${REMOTE_FUELING_PKG}"
+  popd
+fi
 
 # Add partner config.
 PARTNER_CONF=""
@@ -159,10 +164,10 @@ spark-submit \
     --conf spark.kubernetes.executor.request.cores="${EXECUTOR_CORES}" \
 \
     --conf spark.executorEnv.APOLLO_CONDA_ENV="${CONDA_ENV}" \
-    --conf spark.executorEnv.APOLLO_FUELING_PYPATH="${REMOTE_FUELING_PKG}" \
+    --conf spark.executorEnv.APOLLO_FUELING_PYPATH="${FUELING_PKG}" \
     --conf spark.kubernetes.driverEnv.APOLLO_CONDA_ENV="${CONDA_ENV}" \
     --conf spark.kubernetes.driverEnv.APOLLO_EXECUTORS="${EXECUTORS}" \
-    --conf spark.kubernetes.driverEnv.APOLLO_FUELING_PYPATH="${REMOTE_FUELING_PKG}" \
+    --conf spark.kubernetes.driverEnv.APOLLO_FUELING_PYPATH="${FUELING_PKG}" \
     --conf spark.kubernetes.driver.secretKeyRef.APOLLO_EMAIL_PASSWD="apollo-k8s-secret:email-passwd" \
     --conf spark.kubernetes.driver.secretKeyRef.AWS_ACCESS_KEY_ID="bos-secret:ak" \
     --conf spark.kubernetes.driver.secretKeyRef.AWS_SECRET_ACCESS_KEY="bos-secret:sk" \
@@ -174,4 +179,4 @@ spark-submit \
     --conf spark.kubernetes.executor.secretKeyRef.MONGO_PASSWD="mongo-secret:mongo-passwd" \
     ${PARTNER_CONF} \
 \
-    "${REMOTE_JOB_FILE}" --running_mode=PROD $@
+    "${JOB_FILE}" --running_mode=PROD $@
